+++ src/admin/product/AdminCategory.vue
@@ -1,262 +1,200 @@
 <script setup>
-import { ref, reactive, computed } from 'vue'
+import { ref } from 'vue'
 
-// 오늘 날짜
-const today = new Date().toISOString().split("T")[0]
+// [1] 카테고리 관리
+const categories = ref([])
+const isAddingCategory = ref(false)
+const tempCategory = ref('')
+const selectedTopic = ref('') // 주제 선택
 
-// 검색 조건
-const search = reactive({
-  productId: '',
-  productName: '',
-  seller: '',
-})
+// [2] 카테고리 설정
+const categoryId = ref(27)
+const newCategoryName = ref('')
+const showPostCount = ref('표시합니다')
+const showNewPost = ref('표시합니다')
+const newPostDuration = ref('3일')
 
-// 테이블 데이터
-const currentPage = ref(1)
-const pageSize = 10
-
-const products = ref([
-{
-    id: 'P1001',
-    name: '무선 헤드폰',
-    category: '전자제품',
-    price: '150000',
-    stock: '25',
-    registered: '2023-10-01 09:30:00',
-    status: '판매중',
-    overdue: 'N',
-  },
-  {
-    id: 'P1002',
-    name: '청바지',
-    category: '의류',
-    price: '50000',
-    stock: '100',
-    registered: '2023-09-25 14:20:00',
-    status: '판매중',
-    overdue: 'Y',
-  },
-  {
-    id: 'P1003',
-    name: '자바스크립트 입문서',
-    category: '도서',
-    price: '30000',
-    stock: '50',
-    registered: '2023-10-03 11:00:00',
-    status: '판매중',
-    overdue: 'N',
-  },
-  {
-    id: 'P1004',
-    name: '스마트워치',
-    category: '전자제품',
-    price: '200000',
-    stock: '15',
-    registered: '2023-09-28 10:15:00',
-    status: '판매중',
-    overdue: 'N',
-  },
-  {
-    id: 'P1005',
-    name: '아동용 신발',
-    category: '패션',
-    price: '40000',
-    stock: '80',
-    registered: '2023-10-05 12:45:00',
-    status: '판매중',
-    overdue: 'Y',
-  },
-  // 더 추가 가능
-])
-
-// 필터된 결과
-const filteredData = computed(() => {
-  return products.value.filter((item) => {
-    return (
-        item.id.includes(search.productId) &&
-        item.name.includes(search.productName) &&
-        item.price.includes(search.seller)
-    )
-  })
-})
-
-// 페이지네이션
-const paginatedData = computed(() => {
-  const start = (currentPage.value - 1) * pageSize
-  return filteredData.value.slice(start, start + pageSize)
-})
-
-const totalPages = computed(() => {
-  return Math.ceil(filteredData.value.length / pageSize)
-})
-
-// 페이지 이동
-function filterList() {
-  currentPage.value = 1
+// 추가 시작
+const startAdding = () => {
+  isAddingCategory.value = true
+  tempCategory.value = ''
+  selectedTopic.value = ''
 }
 
-function goToPage(page) {
-  currentPage.value = page
+// 추가 확인
+const confirmAdd = () => {
+  const trimmed = tempCategory.value.trim()
+  if (trimmed) {
+    // 주제와 함께 저장
+    categories.value.push({
+      name: trimmed,
+      topic: selectedTopic.value || '주제 없음'
+    })
+    isAddingCategory.value = false
+    tempCategory.value = ''
+    selectedTopic.value = ''
+  } else {
+    alert('추가할 카테고리명을 입력해주세요.')
+  }
+}
+
+// 추가 취소
+const cancelAdd = () => {
+  isAddingCategory.value = false
+  tempCategory.value = ''
+  selectedTopic.value = ''
+}
+
+// 삭제
+const removeCategory = (index) => {
+  categories.value.splice(index, 1)
+}
+
+// 설정 저장
+const saveCategorySettings = () => {
+  alert(`카테고리 설정 저장됨
+  - 글 수: ${showPostCount.value}
+  - 새 글 여부: ${showNewPost.value}
+  - 기준일: ${newPostDuration.value}`)
 }
 </script>
 
 <template>
-  <div class="screen">
-    <div class="root-wrapper">
-      <div class="root">
-        <!-- 바깥 큰 상자 -->
-        <div class="border rounded p-4 bg-light">
+  <div class="border rounded p-4 bg-light test2">
+    <div class="bg-white py-5">
+      <h5 class="fw-bold mb-3">카테고리 관리</h5>
+      <div class="container my-4">
+        <!-- 카테고리 관리 -->
+        <div class="p-4 mb-4 border bg-white">
+          <p class="text-muted">
+            카테고리 순서를 변경하고 하위 연결을 설정할 수 없습니다.
+            <a href="#" class="ms-1 text-danger">자세히 보기</a>
+          </p>
+          <p class="text-muted">드래그 앤 드롭으로 카테고리 순서를 변경할 수 있습니다.</p>
 
-          <!-- 옵션만 따로 감싼 상자 -->
-          <div class="bg-white rounded shadow-sm p-3 mb-3">
-            <div class="d-flex flex-wrap gap-3 align-items-start">
-              <!-- 옵션 폼 그룹 -->
-              <div class="flex-grow-1">
-                <!-- 1행 -->
-                <div class="d-flex flex-wrap align-items-center gap-2 mb-2">
-                  <div class="d-flex align-items-center gap-1">
-                    <label for="inputDate" class="form-label mb-0" >기간</label>
-                    <div class="dropdown">
-                      <button class="btn btn-sm btn-secondary dropdown-toggle" id="inputDate" type="button" data-bs-toggle="dropdown">
-                        상품 등록일
-                      </button>
-                      <ul class="dropdown-menu">
-                        <li><a class="dropdown-item" href="#">상품 등록일</a></li>
-                        <li><a class="dropdown-item" href="#">상품 등록일2</a></li>
-                      </ul>
-                    </div>
-                    <button type="button" class="btn btn-sm btn-primary">전체</button>
-                  </div>
-
-                  <div class="d-flex align-items-center gap-1">
-                    <input type="date" class="form-control form-control-sm" :value="today" style="max-width: 140px;">
-                    <span>~</span>
-                    <input type="date" class="form-control form-control-sm" :value="today" style="max-width: 140px;">
-                  </div>
-
-                  <div class="d-flex align-items-center gap-1 flex-nowrap">
-                    <label class="form-label mb-0 text-nowrap" style="font-size: 12px">검색 구분</label>
-                    <div class="dropdown">
-                      <button class="btn btn-sm btn-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown">
-                        통합 검색
-                      </button>
-                      <ul class="dropdown-menu">
-                        <li><a class="dropdown-item" href="#">상품 등록일</a></li>
-                      </ul>
-                    </div>
-                    <input type="text" class="form-control form-control-sm" placeholder="검색어" style="max-width: 200px;">
-                  </div>
+          <!-- 목록 -->
+          <div class="mb-3">
+            <label class="form-label fw-semibold">분류 전체보기</label>
+            <ul class="list-group">
+              <!-- 기존 카테고리 목록 -->
+              <li
+                v-for="(category, index) in categories"
+                :key="index"
+                class="list-group-item d-flex justify-content-between align-items-center"
+              >
+                <div>
+                  {{ category.name }}
+                  <small class="text-muted ms-2">({{ category.topic }})</small>
                 </div>
+                <button class="btn btn-sm btn-danger" @click="removeCategory(index)">삭제</button>
+              </li>
 
-                <!-- 2행 -->
-                <div class="d-flex flex-wrap align-items-center gap-2">
-                  <div class="d-flex align-items-center gap-1">
-                    <label class="form-label mb-0">창고 선택</label>
-                    <div class="input-group input-group-sm" style="max-width: 200px;">
-                      <input type="text" class="form-control" placeholder="지역 선택">
-                      <button class="btn btn-outline-secondary" type="button">Button</button>
-                    </div>
-                    <div class="input-group input-group-sm" style="max-width: 200px;">
-                      <input type="text" class="form-control" placeholder="창고 선택">
-                      <button class="btn btn-outline-secondary" type="button">Button</button>
-                    </div>
-                  </div>
+              <!-- 추가 입력창 -->
+              <li v-if="isAddingCategory" class="list-group-item d-flex align-items-center flex-wrap gap-2">
+                <input
+                  type="text"
+                  class="form-control"
+                  v-model="tempCategory"
+                  placeholder="카테고리명을 입력하세요"
+                  style="max-width: 300px"
+                />
 
-                  <div class="d-flex align-items-center gap-1">
-                    <label class="form-label mb-0">상품 수량</label>
-                    <div class="dropdown">
-                      <button class="btn btn-sm btn-secondary dropdown-toggle" type="button" data-bs-toggle="dropdown">
-                        현재 상품
-                      </button>
-                      <ul class="dropdown-menu">
-                        <li><a class="dropdown-item" href="#">수리중</a></li>
-                        <li><a class="dropdown-item" href="#">대여중</a></li>
-                        <li><a class="dropdown-item" href="#">가용 재고</a></li>
-                      </ul>
-                    </div>
-                    <input type="text" class="form-control form-control-sm" placeholder="최소 수량" style="max-width: 100px;">
-                    <span>개 이상</span>
-                    <input type="text" class="form-control form-control-sm" placeholder="최대 수량" style="max-width: 100px;">
-                    <span>개 이하</span>
-                  </div>
-                </div>
-              </div>
+                <select class="form-select" v-model="selectedTopic" style="max-width: 200px;">
+                  <option value="">주제 없음</option>
+                  <option>냉장고</option>
+                  <option>세탁기</option>
+                  <option>공기청정기</option>
+                  <option>청소기</option>
+                  <option>전자레인지</option>
+                  <option>커피머신</option>
+                </select>
 
-              <!-- 검색 버튼 (오른쪽) -->
-              <div class="align-self-start">
-                <button type="submit" class="btn btn-primary">검색</button>
-              </div>
+                <button class="btn btn-outline-secondary btn-sm" @click="cancelAdd">취소</button>
+                <button class="btn btn-dark btn-sm" @click="confirmAdd">확인</button>
+              </li>
+
+              <!-- 비어있을 때 -->
+              <li v-if="categories.length === 0 && !isAddingCategory" class="list-group-item text-muted">
+                아직 추가된 카테고리가 없습니다.
+              </li>
+            </ul>
+          </div>
+
+          <!-- 추가 버튼 -->
+          <div
+            class="px-3 py-2 bg-light border rounded d-flex align-items-center mb-3"
+            role="button"
+            @click="startAdding"
+            v-if="!isAddingCategory"
+          >
+            <span class="fw-bold text-danger">+ 카테고리 추가</span>
+            <span class="ms-auto text-muted">{{ categories.length }}/500</span>
+          </div>
+
+          <div class="text-end">
+            <button class="btn btn-dark" @click="confirmAdd">변경사항 저장</button>
+          </div>
+        </div>
+
+        <!-- 카테고리 설정 -->
+        <h5 class="fw-bold mb-3">카테고리 설정</h5>
+        <div class="p-4 border bg-white">
+          <!-- ID -->
+          <div class="row mb-3">
+            <label class="col-sm-2 col-form-label fw-semibold">카테고리 ID</label>
+            <div class="col-sm-10 d-flex align-items-center">
+              <input type="text" class="form-control me-2" v-model="categoryId" style="max-width: 100px" />
+              <span>글자까지 표시됩니다.</span>
             </div>
           </div>
 
-          <!-- 🔸 가운데 정렬된 제목 -->
+          <!-- 글 수 표시 여부 -->
+          <div class="row mb-3">
+            <label class="col-sm-2 col-form-label fw-semibold">카테고리별 글 수</label>
+            <div class="col-sm-10">
+              <select class="form-select" v-model="showPostCount" style="max-width: 300px;">
+                <option>표시합니다</option>
+                <option>표시하지 않습니다</option>
+              </select>
+            </div>
+          </div>
 
-        </div>
-        <div class="p-3">
+          <!-- 새 글 여부 표시 -->
+          <div class="row mb-3">
+            <label class="col-sm-2 col-form-label fw-semibold">카테고리에 새 글 발생 여부</label>
+            <div class="col-sm-10">
+              <select class="form-select" v-model="showNewPost" style="max-width: 300px;">
+                <option>표시합니다</option>
+                <option>표시하지 않습니다</option>
+              </select>
+            </div>
+          </div>
 
-        <div class="text-center border-top pt-3 mt-3 mb-2">
-          <h5 class="fw-bold mb-0">상품관리</h5>
-        </div>
-        <!-- 테이블 -->
-        <table class="table table-bordered table-hover text-center">
-          <thead class="custom-thead" >
-          <tr>
-            <th>상품 ID</th>
-            <th>상품명</th>
-            <th>카테고리</th>
-            <th>가격</th>
-            <th>재고 수량</th>
-            <th>등록일</th>
-            <th>판매 상태</th>
-            <th>연체 여부</th>
-          </tr>
-          </thead>
-          <tbody>
-          <tr v-for="item in paginatedData" :key="item.id">
-            <td>{{ item.id }}</td>
-                <td>{{ item.name }}</td>
-                <td>{{ item.category }}</td>
-                <td>{{ item.price }}</td>
-                <td>{{ item.stock }}</td>
-                <td>{{ item.registered }}</td>
-                <td>{{ item.status }}</td>
-                <td>{{ item.overdue }}</td>
-          </tr>
-          </tbody>
-        </table>
+          <!-- 조건 드롭다운 -->
+          <div class="row mb-3" v-if="showNewPost === '표시합니다'">
+            <label class="col-sm-2 col-form-label fw-semibold">·</label>
+            <div class="col-sm-10 d-flex align-items-center">
+              <select class="form-select me-2" v-model="newPostDuration" style="width: 100px;">
+                <option>1일</option>
+                <option>3일</option>
+                <option>5일</option>
+              </select>
+              <span class="text-muted">이내에 발생된 글이 있으면 새 글 여부를 표시합니다.</span>
+            </div>
+          </div>
 
-        <!-- 페이지네이션 -->
-        <nav class="d-flex justify-content-center">
-          <ul class="pagination">
-            <li class="page-item" :class="{ disabled: currentPage === 1 }">
-              <a class="page-link" href="#" @click.prevent="currentPage--">‹</a>
-            </li>
-            <li class="page-item" v-for="page in totalPages" :key="page" :class="{ active: page === currentPage }">
-              <a class="page-link" href="#" @click.prevent="goToPage(page)">{{ page }}</a>
-            </li>
-            <li class="page-item" :class="{ disabled: currentPage === totalPages }">
-              <a class="page-link" href="#" @click.prevent="currentPage++">›</a>
-            </li>
-          </ul>
-        </nav>
-
+          <!-- 저장 버튼 -->
+          <div class="text-end">
+            <button class="btn btn-dark" @click="saveCategorySettings">변경사항 저장</button>
+          </div>
         </div>
       </div>
     </div>
   </div>
 </template>
 
-
-
 <style scoped>
-.root {
-  font-size: 12px
-}
-.table {
-  font-size: 14px;
-}
-
-.custom-thead th {
-  background-color: #DFF0D8 !important;
+.test2 {
+  width: 990.4px;
 }
 </style>
